+++ packages/server/src/modules/friend/entities/friendGroupMember.entity.ts
@@ -4,7 +4,7 @@ import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
  * friend group member entity
  */
 @Entity('friend_group_member')
-export class FriendGroupMember {
+export class FriendGroupMemberEntity {
   @PrimaryGeneratedColumn()
   id: number;