+++ packages/server/src/modules/friend/friend.module.ts
@@ -1,8 +1,11 @@
 import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { FriendEntity } from './entities/friend.entity';
 import { FriendController } from './friend.controller';
 import { FriendService } from './friend.service';
 
 @Module({
+  imports: [TypeOrmModule.forFeature([FriendEntity])],
   controllers: [FriendController],
   providers: [FriendService],
 })