+++ packages/server/src/modules/user/user.dto.ts
@@ -1,0 +1,6 @@
+import { ApiProperty } from '@nestjs/swagger';
+
+export class SearchUserDTO {
+  @ApiProperty({ description: '用户 id' })
+  key: string;
+}